Show survey prompt with colors
